finished the setup of the package
