+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include LICENSE
+recursive-include Querio/lib *
