+++ Querio/storage/csed25/Flights/Flights_schema.json
@@ -1,0 +1,1 @@
+{"__path__": "C:\\Users\\pc\\OneDrive-AlexandriaUniversity\\Documents\\GitHub\\SimpleFSDB\\Querio\\storage\\csed25\\Flights", "name": "Flights", "primary_key": "FlightId", "columns": ["FlightId", "PlaneId", "Seats_reserved", "Take_off", "Destination", "Stops", "Departure_time", "Departure_date", "Arrival_time", "Arrival_date"], "overwrite": "False", "index_keys": ["FlightId", "PlaneId", "Seats_reserved"]}
