+++ Querio/storage/csed25/Reservations/Reservations_schema.json
@@ -1,0 +1,1 @@
+{"__path__": "C:\\Users\\pc\\OneDrive-AlexandriaUniversity\\Documents\\GitHub\\SimpleFSDB\\Querio\\storage\\csed25\\Reservations", "name": "Reservations", "primary_key": "ReservationId", "columns": ["ReservationId", "First_name", "Last_name", "Passport_Id", "Gender", "Age", "Flight_Id", "Luggage_allowance", "Seat"], "overwrite": "False", "index_keys": ["ReservationId", "First_name", "Last_name", "Seat"]}
