--- pytestexample/requirments.txt
@@ -1,9 +1,0 @@
-attrs==22.1.0
-colorama==0.4.5
-iniconfig==1.1.1
-packaging==21.3
-pluggy==1.0.0
-py==1.11.0
-pyparsing==3.0.9
-pytest==7.1.3
-tomli==2.0.1
