--- pytestexample/venv.env
@@ -1,1 +1,0 @@
-ECHO is on.
