--- src/commands/create_command.py
@@ -1,43 +1,0 @@
-# Import necessary modules
-import sys
-import os
-
-# Append the parent directory of the current file to sys.path
-sys.path.append(os.path.dirname(__file__).replace("commands", ''))
-
-# Import custom classes
-from commands.abstract_command import *
-from model.database import *
-
-
-# Define a CreateCommand class that inherits from AbstractCommand
-class CreateCommand(AbstractCommand):
-    def __init__(self, schema_path):
-        # Validate the schema_path parameter
-        CreateCommand.__validate__(schema_path)
-
-        # Store the schema_path and retrieve the schema data
-        self.schema_path = schema_path
-        self.schema_data = CreateCommand.__get_database_schema__(schema_path)
-
-    # Execute method to create a database based on the schema
-    def execute(self):
-        # Create a Database instance using the schema data
-        database = Database(self.schema_data)
-        # Serialize and store the database
-
-    @staticmethod
-    def __validate__(schema_path):
-        # Check if the schema_path is empty or None
-        if not schema_path:
-            raise NoParameterError("schema_path parameter not entered")
-
-        # Check if the schema file exists
-        if not os.path.isfile(schema_path):
-            raise WrongParameterError("Schema doesn't exist")
-
-    @staticmethod
-    def __get_database_schema__(schema_path):
-        # Read and load the schema data from the specified file
-        with open(schema_path, 'r') as file:
-            return json.load(file)
