--- src/input_adaptors/iadaptor.py
@@ -1,6 +1,0 @@
-# Import the `ABC` (Abstract Base Class) from the `abc` module
-from abc import ABC
-
-# Define an abstract base class called `IInputAdaptor` that inherits from `ABC`
-class IInputAdaptor(ABC):
-    pass
